chore: added wip home-chess-demo story
