+++ apps/storybook/tsconfig.json
@@ -8,6 +8,7 @@
 			"@pluv-apps/server-cloudflare": ["../server-cloudflare/src"],
 			"@pluv-apps/web": ["../web/src"],
 			"@pluv-apps/web/components": ["../web/components"],
+			"@pluv-apps/web/pluv-io": ["../web/pluv-io"],
 			"@pluv-internal/cloudflare-utils": ["../../internal/cloudflare-utils/dist"],
 			"@pluv-internal/rainfall": ["../../internal/rainfall/dist"],
 			"@pluv-internal/react-hooks": ["../../internal/react-hooks/dist"],
