+++ apps/web/package.json
@@ -9,7 +9,8 @@
         "lint": "next lint --fix"
     },
     "files": [
-        "components.ts"
+        "components.ts",
+        "pluv-io.ts"
     ],
     "dependencies": {
         "@pluv-apps/server-cloudflare": "workspace:^",
