+++ apps/web/src/components/HomeChessDemo/HomeChessDemo.tsx
@@ -1,6 +1,7 @@
-import type { CSSProperties, FC } from "react";
-import { PluvRoomProvider } from "../../pluv-io/cloudflare";
-import { HomeChessDemoBoard } from "./HomeChessDemoBoard";
+import { ChessBoard } from "@pluv-internal/react-chess";
+import { y } from "@pluv/react";
+import { CSSProperties, FC, useMemo } from "react";
+import { usePluvStorage } from "../../pluv-io/cloudflare";
 
 export interface HomeChessDemoProps {
     className?: string;
@@ -8,9 +9,18 @@ export interface HomeChessDemoProps {
 }
 
 export const HomeChessDemo: FC<HomeChessDemoProps> = ({ className, style }) => {
+    const [demo, sharedType] = usePluvStorage("demo");
+
+    const history = useMemo(() => demo?.chessHistory?.slice() ?? [], [demo]);
+
     return (
-        <PluvRoomProvider room="home-chess">
-            <HomeChessDemoBoard className={className} style={style} />
-        </PluvRoomProvider>
+        <ChessBoard
+            className={className}
+            history={history}
+            onMove={({ game }) => {
+                sharedType?.set("chessHistory", y.array(game.history()));
+            }}
+            style={style}
+        />
     );
 };
